refactor: better import map and folder structure
