+++ import_map.json
@@ -1,10 +1,10 @@
 {
 	"imports": {
 		"@/": "./src/",
-		"std/collection/": "https://deno.land/std@0.152.0/collections/",
-		"std/path": "https://deno.land/std@0.152.0/path/mod.ts",
-		"std/colors": "https://deno.land/std@0.152.0/fmt/colors.ts",
-		"std/time": "https://deno.land/std@0.152.0/datetime/mod.ts",
-		"std/fs": "https://deno.land/std@0.152.0/fs/mod.ts"
+		"collection/": "https://deno.land/std@0.152.0/collections/",
+		"fmt/": "https://deno.land/std@0.152.0/fmt/",
+		"path": "https://deno.land/std@0.152.0/path/mod.ts",
+		"datetime": "https://deno.land/std@0.152.0/datetime/mod.ts",
+		"fs": "https://deno.land/std@0.152.0/fs/mod.ts"
 	}
 }
